saturn_dw/settings/local.py
from .base import *
from dotenv import load_dotenv
import os

# Load .env file
load_dotenv(dotenv_path=BASE_DIR / ".env")

# Database config from env
DATABASES = {
    "default": {
        "ENGINE": "django.db.backends.postgresql",
        "NAME": os.getenv("DB_NAME", "saturn_db"),
        "USER": os.getenv("DB_USER", "postgres"),
        "PASSWORD": os.getenv("DB_PASSWORD", "postgres"),
        "HOST": os.getenv("DB_HOST", "localhost"),
        "PORT": os.getenv("DB_PORT", "5432"),
    }
}
