requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==4.2.23
gunicorn==23.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
